+++ emoji/unicode_codes/__init__.py
@@ -1,11 +1,25 @@
-from typing import Optional
+import json
 from functools import lru_cache
-from emoji.unicode_codes.data_dict import EMOJI_DATA, STATUS, LANGUAGES
+from pathlib import Path
+from warnings import warn
+
+from typing import Optional, Dict, List, Any
+
+from emoji.unicode_codes.data_dict import STATUS, LANGUAGES
 
 __all__ = [
     'get_emoji_by_name',
-    'EMOJI_DATA', 'STATUS', 'LANGUAGES'
+    'load_from_json',
+    'EMOJI_DATA',
+    'STATUS',
+    'LANGUAGES',
 ]
+
+_DEFAULT_KEYS = ('en', 'alias', 'E', 'status')  # The keys in emoji.json
+
+_loaded_keys: List[str] = list(
+    _DEFAULT_KEYS
+)  # Keep track of keys already loaded from json files to avoid loading them twice
 
 
 @lru_cache(maxsize=4000)
@@ -31,3 +45,61 @@ def get_emoji_by_name(name: str, language: str) -> Optional[str]:
             return emj
 
     return None
+
+
+class EmojiDataDict(Dict[str, Any]):
+    """Replaces built-in-dict in the values of the EMOJI_DATA dict.
+    Auto loads language data when accessing language data via
+    key-access without prior loading of the language:
+    e.g. EMOJI_DATA['👌']['fr'] will auto load French language and not throw
+    a KeyError.
+    Shows a deprecation warning explainging that `emoji.config.load_language()`
+    should be used."""
+
+    def __missing__(self, key: str) -> str:
+        """Auto load language `key`, raises KeyError if language is no supported."""
+        if key in LANGUAGES and key not in _loaded_keys:
+            load_from_json(key)
+            if key in self:
+                warn(
+                    f"""Use emoji.config.load_language('{key}') before accesing EMOJI_DATA[emj]['{key}'].
+Accessing EMOJI_DATA[emj]['{key}'] without loading the language is deprecated.""",
+                    DeprecationWarning,
+                    stacklevel=3,
+                )
+                return self[key]  # type: ignore
+
+        raise KeyError(key)
+
+
+EMOJI_DATA: Dict[str, Dict[str, Any]]
+
+
+def _load_default_from_json():
+    global EMOJI_DATA
+    global _loaded_keys
+
+    file = Path(__file__).with_name('emoji.json')
+    with open(file, 'rb') as f:
+        EMOJI_DATA = dict(json.load(f, object_pairs_hook=EmojiDataDict))  # type: ignore
+    _loaded_keys = list(_DEFAULT_KEYS)
+
+
+def load_from_json(key: str):
+    """Load values from the file 'emoji_{key}.json' into EMOJI_DATA"""
+
+    if key in _loaded_keys:
+        return
+
+    if key not in LANGUAGES:
+        raise NotImplementedError('Language not supported', key)
+
+    file = Path(__file__).with_name(f'emoji_{key}.json')
+    with open(file, 'rb') as f:
+        for emj, value in json.load(f).items():
+            EMOJI_DATA[emj][key] = value  # type: ignore
+
+    _loaded_keys.append(key)
+
+
+_load_default_from_json()