+++ tests/test_core.py
@@ -6,43 +6,62 @@ from typing import Any, Callable, Dict, List, Tuple, Union
 from typing_extensions import Literal
 import pytest
 import emoji.unicode_codes
-from testutils import ascii, normalize, all_language_packs, all_language_and_alias_packs, get_emoji_unicode_dict
+from testutils import (
+    ascii,
+    normalize,
+    all_language_packs,
+    all_language_and_alias_packs,
+    get_emoji_unicode_dict,
+    load_all_languages as load_all_languages,
+)
+
+try:
+    from typeguard import suppress_type_checks  # type:ignore
+except ImportError:
+    from contextlib import nullcontext as suppress_type_checks  # type:ignore
 
 
-def test_emojize_name_only():
+def test_emojize_name_only(load_all_languages):  # type:ignore
     # Check that the regular expression emoji.core._EMOJI_NAME_PATTERN contains all the necesseary characters
     from emoji.core import _EMOJI_NAME_PATTERN  # pyright: ignore [reportPrivateUsage]
 
-    pattern = re.compile('[^%s]' % (_EMOJI_NAME_PATTERN, ))
+    pattern = re.compile('[^%s]' % (_EMOJI_NAME_PATTERN,))
 
     for lang_code, emoji_pack in all_language_and_alias_packs():
         for name_in_db in emoji_pack.keys():
-
             pairs = [
                 ('Form from EMOJI_DATA', name_in_db),
                 ('NFKC', normalize('NFKC', name_in_db)),
                 ('NFKD', normalize('NFKD', name_in_db)),
                 ('NFD', normalize('NFD', name_in_db)),
-                ('NFC', normalize('NFC', name_in_db))
+                ('NFC', normalize('NFC', name_in_db)),
             ]
             for form, name in pairs:
                 actual = emoji.emojize(name, language=lang_code)
                 expected = emoji_pack[name_in_db]
 
                 if expected != actual:
-                    print("Regular expression is missing a character:")
-                    print("Emoji name %r in form %r contains:" % (name, form))
-                    print("\n".join(["%r (%r) is not in the regular expression" % (x, x.encode('unicode-escape').decode()) for x in pattern.findall(name[1:-1])]))
+                    print('Regular expression is missing a character:')
+                    print('Emoji name %r in form %r contains:' % (name, form))
+                    print(
+                        '\n'.join(
+                            [
+                                '%r (%r) is not in the regular expression'
+                                % (x, x.encode('unicode-escape').decode())
+                                for x in pattern.findall(name[1:-1])
+                            ]
+                        )
+                    )
 
                 assert expected == actual, '%s != %s' % (expected, actual)
                 assert pattern.search(name[1:-1]) is None
 
 
-def test_regular_expression_minimal():
+def test_regular_expression_minimal(load_all_languages):  # type:ignore
     # Check that the regular expression emoji.core._EMOJI_NAME_PATTERN only contains the necesseary characters
     from emoji.core import _EMOJI_NAME_PATTERN  # pyright: ignore [reportPrivateUsage]
 
-    pattern_str = '[^%s]' % (_EMOJI_NAME_PATTERN, )
+    pattern_str = '[^%s]' % (_EMOJI_NAME_PATTERN,)
     i = 2
     while i < len(pattern_str) - 1:
         c = pattern_str[i]
@@ -59,7 +78,7 @@ def test_regular_expression_minimal():
                     normalize('NFKC', name_in_db),
                     normalize('NFKD', name_in_db),
                     normalize('NFD', name_in_db),
-                    normalize('NFC', name_in_db)
+                    normalize('NFC', name_in_db),
                 ]
                 for str in names:
                     if pattern.search(str):
@@ -68,7 +87,7 @@ def test_regular_expression_minimal():
             if failed:
                 break
         if not failed:
-            assert failed, "char: %r is not necessary in regular expression" % (c, )
+            assert failed, 'char: %r is not necessary in regular expression' % (c,)
 
         i += 1
 
@@ -76,12 +95,12 @@ def test_regular_expression_minimal():
 def test_emojize_complicated_string():
     # A bunch of emoji's with UTF-8 strings to make sure the regex expression is functioning
     name_code = {
-        ':flag_for_Ceuta_&_Melilla:': '\U0001F1EA\U0001F1E6',
-        ':flag_for_St._Barthélemy:': '\U0001F1E7\U0001F1F1',
-        ':flag_for_Côte_d’Ivoire:': '\U0001F1E8\U0001F1EE',
-        ':flag_for_Åland_Islands:': '\U0001F1E6\U0001F1FD',
-        ':flag_for_São_Tomé_&_Príncipe:': '\U0001F1F8\U0001F1F9',
-        ':flag_for_Curaçao:': '\U0001F1E8\U0001F1FC'
+        ':flag_for_Ceuta_&_Melilla:': '\U0001f1ea\U0001f1e6',
+        ':flag_for_St._Barthélemy:': '\U0001f1e7\U0001f1f1',
+        ':flag_for_Côte_d’Ivoire:': '\U0001f1e8\U0001f1ee',
+        ':flag_for_Åland_Islands:': '\U0001f1e6\U0001f1fd',
+        ':flag_for_São_Tomé_&_Príncipe:': '\U0001f1f8\U0001f1f9',
+        ':flag_for_Curaçao:': '\U0001f1e8\U0001f1fc',
     }
     string = ' complicated! '.join(list(name_code.keys()))
     actual = emoji.emojize(string)
@@ -92,13 +111,13 @@ def test_emojize_complicated_string():
     assert expected == actual, '%s != %s' % (expected, actual)
 
 
-def test_emojize_languages():
+def test_emojize_languages(load_all_languages):  # type:ignore
     for lang_code, emoji_pack in all_language_packs():
         for name, emj in emoji_pack.items():
             assert emoji.emojize(name, language=lang_code) == emj
 
 
-def test_demojize_languages():
+def test_demojize_languages(load_all_languages):  # type:ignore
     for lang_code, emoji_pack in all_language_packs():
         for name, emj in emoji_pack.items():
             assert emoji.demojize(emj, language=lang_code) == name
@@ -110,48 +129,66 @@ def test_emojize_variant():
 
     english_pack = get_emoji_unicode_dict('en')
 
-    assert emoji.emojize(
-        ':Taurus:', variant=None) == english_pack[':Taurus:']
+    assert emoji.emojize(':Taurus:', variant=None) == english_pack[':Taurus:']
     assert emoji.emojize(':Taurus:', variant=None) == emoji.emojize(':Taurus:')
-    assert emoji.emojize(':Taurus:', variant='text_type') == remove_variant(
-        english_pack[':Taurus:']) + '\ufe0e'
-    assert emoji.emojize(':Taurus:', variant='emoji_type') == remove_variant(
-        english_pack[':Taurus:']) + '\ufe0f'
+    assert (
+        emoji.emojize(':Taurus:', variant='text_type')
+        == remove_variant(english_pack[':Taurus:']) + '\ufe0e'
+    )
+    assert (
+        emoji.emojize(':Taurus:', variant='emoji_type')
+        == remove_variant(english_pack[':Taurus:']) + '\ufe0f'
+    )
 
-    assert emoji.emojize(
-        ':admission_tickets:', variant=None) == english_pack[':admission_tickets:']
+    assert (
+        emoji.emojize(':admission_tickets:', variant=None)
+        == english_pack[':admission_tickets:']
+    )
     assert emoji.emojize(':admission_tickets:', variant=None) == emoji.emojize(
-        ':admission_tickets:')
-    assert emoji.emojize(':admission_tickets:', variant='text_type') == remove_variant(
-        english_pack[':admission_tickets:']) + '\ufe0e'
-    assert emoji.emojize(':admission_tickets:', variant='emoji_type') == remove_variant(
-        english_pack[':admission_tickets:']) + '\ufe0f'
+        ':admission_tickets:'
+    )
+    assert (
+        emoji.emojize(':admission_tickets:', variant='text_type')
+        == remove_variant(english_pack[':admission_tickets:']) + '\ufe0e'
+    )
+    assert (
+        emoji.emojize(':admission_tickets:', variant='emoji_type')
+        == remove_variant(english_pack[':admission_tickets:']) + '\ufe0f'
+    )
 
-    with pytest.raises(ValueError):
-        emoji.emojize(':admission_tickets:', variant=False)  # type: ignore[arg-type]
+    with suppress_type_checks():  # type:ignore
+        with pytest.raises(ValueError):
+            emoji.emojize(':admission_tickets:', variant=False)  # type: ignore[arg-type]
 
-    with pytest.raises(ValueError):
-        emoji.emojize(':admission_tickets:', variant=True)  # type: ignore[arg-type]
+        with pytest.raises(ValueError):
+            emoji.emojize(':admission_tickets:', variant=True)  # type: ignore[arg-type]
 
-    with pytest.raises(ValueError):
-        emoji.emojize(':admission_tickets:', variant='wrong')  # type: ignore[arg-type]
+        with pytest.raises(ValueError):
+            emoji.emojize(':admission_tickets:', variant='wrong')  # type: ignore[arg-type]
 
-    assert emoji.emojize(":football:") == ':football:'
-    assert emoji.emojize(":football:", variant="text_type") == ':football:'
-    assert emoji.emojize(":football:", language="alias") == '\U0001F3C8'
-    assert emoji.emojize(":football:", variant="emoji_type", language="alias") == '\U0001F3C8'
+    assert emoji.emojize(':football:') == ':football:'
+    assert emoji.emojize(':football:', variant='text_type') == ':football:'
+    assert emoji.emojize(':football:', language='alias') == '\U0001f3c8'
+    assert (
+        emoji.emojize(':football:', variant='emoji_type', language='alias')
+        == '\U0001f3c8'
+    )
 
 
 def test_demojize_removes_variant():
     # demojize should remove all variant indicators \ufe0e and \ufe0f from the string
-    text = "".join([emoji.emojize(':Taurus:', variant='text_type'),
-                    emoji.emojize(':Taurus:', variant='emoji_type'),
-                    emoji.emojize(':admission_tickets:', variant='text_type'),
-                    emoji.emojize(':admission_tickets:', variant='emoji_type'),
-                    emoji.emojize(':alien:', variant='text_type'),
-                    emoji.emojize(':alien:', variant='emoji_type'),
-                    emoji.emojize(':atom_symbol:', variant='text_type'),
-                    emoji.emojize(':atom_symbol:', variant='emoji_type')])
+    text = ''.join(
+        [
+            emoji.emojize(':Taurus:', variant='text_type'),
+            emoji.emojize(':Taurus:', variant='emoji_type'),
+            emoji.emojize(':admission_tickets:', variant='text_type'),
+            emoji.emojize(':admission_tickets:', variant='emoji_type'),
+            emoji.emojize(':alien:', variant='text_type'),
+            emoji.emojize(':alien:', variant='emoji_type'),
+            emoji.emojize(':atom_symbol:', variant='text_type'),
+            emoji.emojize(':atom_symbol:', variant='emoji_type'),
+        ]
+    )
 
     for lang_code in emoji.LANGUAGES:
         result = emoji.demojize(text, language=lang_code)
@@ -167,104 +204,171 @@ def test_emojize_invalid_emoji():
     assert emoji.emojize(string) == string
 
 
-def test_emojize_version():
-    assert emoji.emojize("Flags like :Belgium: are in version 2.0", version=1.0) == "Flags like  are in version 2.0"
-    assert emoji.emojize("Flags like :Belgium: are in version 2.0", version=1.9) == "Flags like  are in version 2.0"
-    assert emoji.emojize("Flags like :Belgium: are in version 2.0", version=2.0) == "Flags like 🇧🇪 are in version 2.0"
-    assert emoji.emojize("Flags like :Belgium: are in version 2.0", version=3.0) == "Flags like 🇧🇪 are in version 2.0"
-    assert emoji.emojize("Boxing gloves :boxing_glove: are in version 3.0", version=0) == "Boxing gloves  are in version 3.0"
-    assert emoji.emojize("Boxing gloves :boxing_glove: are in version 3.0", version=1) == "Boxing gloves  are in version 3.0"
-    assert emoji.emojize("Boxing gloves :boxing_glove: are in version 3.0", version=2) == "Boxing gloves  are in version 3.0"
-    assert emoji.emojize("Boxing gloves :boxing_glove: are in version 3.0", version=3) == "Boxing gloves 🥊 are in version 3.0"
-    assert emoji.emojize("Boxing gloves :boxing_glove: are in version 3.0", version=4) == "Boxing gloves 🥊 are in version 3.0"
+def test_emojize_version(load_all_languages):  # type:ignore
+    assert (
+        emoji.emojize('Flags like :Belgium: are in version 2.0', version=1.0)
+        == 'Flags like  are in version 2.0'
+    )
+    assert (
+        emoji.emojize('Flags like :Belgium: are in version 2.0', version=1.9)
+        == 'Flags like  are in version 2.0'
+    )
+    assert (
+        emoji.emojize('Flags like :Belgium: are in version 2.0', version=2.0)
+        == 'Flags like 🇧🇪 are in version 2.0'
+    )
+    assert (
+        emoji.emojize('Flags like :Belgium: are in version 2.0', version=3.0)
+        == 'Flags like 🇧🇪 are in version 2.0'
+    )
+    assert (
+        emoji.emojize('Boxing gloves :boxing_glove: are in version 3.0', version=0)
+        == 'Boxing gloves  are in version 3.0'
+    )
+    assert (
+        emoji.emojize('Boxing gloves :boxing_glove: are in version 3.0', version=1)
+        == 'Boxing gloves  are in version 3.0'
+    )
+    assert (
+        emoji.emojize('Boxing gloves :boxing_glove: are in version 3.0', version=2)
+        == 'Boxing gloves  are in version 3.0'
+    )
+    assert (
+        emoji.emojize('Boxing gloves :boxing_glove: are in version 3.0', version=3)
+        == 'Boxing gloves 🥊 are in version 3.0'
+    )
+    assert (
+        emoji.emojize('Boxing gloves :boxing_glove: are in version 3.0', version=4)
+        == 'Boxing gloves 🥊 are in version 3.0'
+    )
 
-    assert emoji.emojize("Biking :man_biking: is in 4.0", version=3.0, handle_version='<emoji>') == "Biking <emoji> is in 4.0"
-    assert emoji.emojize("Biking :man_biking: is in 4.0", version=3.0, handle_version=lambda e, data: '<emoji>') == "Biking <emoji> is in 4.0"
-    assert emoji.emojize("Biking :man_biking: is in 4.0", version=3.0, handle_version=lambda e, data: data["fr"]) == "Biking :cycliste_homme: is in 4.0"
+    assert (
+        emoji.emojize(
+            'Biking :man_biking: is in 4.0', version=3.0, handle_version='<emoji>'
+        )
+        == 'Biking <emoji> is in 4.0'
+    )
+    assert (
+        emoji.emojize(
+            'Biking :man_biking: is in 4.0',
+            version=3.0,
+            handle_version=lambda e, data: '<emoji>',
+        )
+        == 'Biking <emoji> is in 4.0'
+    )
+    assert (
+        emoji.emojize(
+            'Biking :man_biking: is in 4.0',
+            version=3.0,
+            handle_version=lambda e, data: data['fr'],
+        )
+        == 'Biking :cycliste_homme: is in 4.0'
+    )
 
     def f(emj: str, data: Dict[str, str]) -> str:
         assert data['E'] == 5
         return ''
 
     assert emoji.emojize(':bowl_with_spoon:', version=-1, handle_version=f) == ''
-    assert emoji.emojize(':bowl_with_spoon:') == '\U0001F963'
+    assert emoji.emojize(':bowl_with_spoon:') == '\U0001f963'
     assert emoji.emojize(':bowl_with_spoon:', version=4) == ''
     assert emoji.emojize(':bowl_with_spoon:', version=4.9) == ''
-    assert emoji.emojize(':bowl_with_spoon:', version=5) == '\U0001F963'
-    assert emoji.emojize(':bowl_with_spoon:', version=5.1) == '\U0001F963'
-    assert emoji.emojize(':bowl_with_spoon:', version=6) == '\U0001F963'
-    assert emoji.emojize(':bowl_with_spoon:', version=4,
-                         handle_version='abc') == 'abc'
-    assert emoji.emojize(':bowl_with_spoon:', version=4,
-                         handle_version=None) == ''
-    assert emoji.emojize(':bowl_with_spoon:', version=4,
-                         handle_version='') == ''
-    assert emoji.emojize(':bowl_with_spoon:', version=4,
-                         handle_version=lambda e, d: str(d['E'])) == '5'
+    assert emoji.emojize(':bowl_with_spoon:', version=5) == '\U0001f963'
+    assert emoji.emojize(':bowl_with_spoon:', version=5.1) == '\U0001f963'
+    assert emoji.emojize(':bowl_with_spoon:', version=6) == '\U0001f963'
+    assert emoji.emojize(':bowl_with_spoon:', version=4, handle_version='abc') == 'abc'
+    assert emoji.emojize(':bowl_with_spoon:', version=4, handle_version=None) == ''
+    assert emoji.emojize(':bowl_with_spoon:', version=4, handle_version='') == ''
+    assert (
+        emoji.emojize(
+            ':bowl_with_spoon:', version=4, handle_version=lambda e, d: str(d['E'])
+        )
+        == '5'
+    )
 
 
 def test_demojize_version():
+    assert (
+        emoji.emojize('A :T-Rex: is eating a :croissant:', version=3.0)
+        == 'A  is eating a 🥐'
+    )
 
-    assert emoji.emojize('A :T-Rex: is eating a :croissant:', version=3.0) == 'A  is eating a 🥐'
+    assert (
+        emoji.emojize(
+            'A :T-Rex: is eating a :croissant:',
+            version=3.0,
+            handle_version='[Unsupported emoji]',
+        )
+        == 'A [Unsupported emoji] is eating a 🥐'
+    )
 
-    assert emoji.emojize('A :T-Rex: is eating a :croissant:', version=3.0, handle_version='[Unsupported emoji]') == 'A [Unsupported emoji] is eating a 🥐'
+    assert (
+        emoji.demojize('A 🦖 is eating a 🥐', version=3.0)
+        == 'A  is eating a :croissant:'
+    )
 
-    assert emoji.demojize('A 🦖 is eating a 🥐', version=3.0) == 'A  is eating a :croissant:'
+    assert (
+        emoji.demojize('A 🦖 is eating a 🥐', handle_version='X', version=3.0)
+        == 'A X is eating a :croissant:'
+    )
 
-    assert emoji.demojize('A 🦖 is eating a 🥐', handle_version='X', version=3.0) == 'A X is eating a :croissant:'
-
-    assert emoji.demojize('A 🦖 is eating a 🥐', handle_version='X', version=5.0) == 'A :T-Rex: is eating a :croissant:'
+    assert (
+        emoji.demojize('A 🦖 is eating a 🥐', handle_version='X', version=5.0)
+        == 'A :T-Rex: is eating a :croissant:'
+    )
 
 
 def test_alias():
     # When lanugage != "alias" aliases should be passed through untouched
     assert emoji.emojize(':soccer:') == ':soccer:'
-    assert emoji.emojize(':soccer:', language="alias") == '\U000026BD'
+    assert emoji.emojize(':soccer:', language='alias') == '\U000026bd'
     assert emoji.emojize(':football:') == ':football:'
-    assert emoji.emojize(':football:', language="alias") == '\U0001F3C8'
+    assert emoji.emojize(':football:', language='alias') == '\U0001f3c8'
     # Multiple aliases for one emoji:
-    assert emoji.emojize(':thumbsup:', language="alias") == emoji.emojize(
-        ':+1:', language="alias")
-    assert emoji.emojize(':thumbsup:', language="alias") == emoji.emojize(
-        ':thumbs_up:', language="alias")
-    assert emoji.emojize(':thumbsup:', language="alias") == '\U0001f44d'
+    assert emoji.emojize(':thumbsup:', language='alias') == emoji.emojize(
+        ':+1:', language='alias'
+    )
+    assert emoji.emojize(':thumbsup:', language='alias') == emoji.emojize(
+        ':thumbs_up:', language='alias'
+    )
+    assert emoji.emojize(':thumbsup:', language='alias') == '\U0001f44d'
 
     thumbsup = '\U0001f44d'
-    assert emoji.demojize(thumbsup, language="alias") != thumbsup
-    assert emoji.demojize(thumbsup, language="alias") != ':thumbs_up:'
-    assert emoji.demojize(thumbsup, language="alias") != emoji.demojize(thumbsup)
+    assert emoji.demojize(thumbsup, language='alias') != thumbsup
+    assert emoji.demojize(thumbsup, language='alias') != ':thumbs_up:'
+    assert emoji.demojize(thumbsup, language='alias') != emoji.demojize(thumbsup)
 
     thailand = '🇹🇭'
-    assert emoji.demojize(thailand, language="alias") != thailand
-    assert emoji.demojize(thailand, language="alias") != ':Thailand:'
-    assert emoji.demojize(thailand, language="alias") != emoji.demojize(thailand)
-    assert emoji.demojize(thailand, language="alias", version=1.0) != emoji.demojize(
-        thailand, language="alias")
+    assert emoji.demojize(thailand, language='alias') != thailand
+    assert emoji.demojize(thailand, language='alias') != ':Thailand:'
+    assert emoji.demojize(thailand, language='alias') != emoji.demojize(thailand)
+    assert emoji.demojize(thailand, language='alias', version=1.0) != emoji.demojize(
+        thailand, language='alias'
+    )
 
     # No alias
     for emj, emoji_data in emoji.EMOJI_DATA.items():
         if emoji_data['status'] != emoji.STATUS['fully_qualified']:
             continue
         if 'alias' not in emoji_data:
-            assert emoji.emojize(emoji_data['en'], language="alias") != emoji_data['en']
-            assert emoji.demojize(emj, language="alias") == emoji_data['en']
+            assert emoji.emojize(emoji_data['en'], language='alias') != emoji_data['en']
+            assert emoji.demojize(emj, language='alias') == emoji_data['en']
 
 
 def test_invalid_alias():
     # Invalid aliases should be passed through untouched
-    assert emoji.emojize(':tester:', language="alias") == ':tester:'
-    assert emoji.emojize(':footbal:', language="alias") == ':footbal:'
-    assert emoji.emojize(':socer:', language="alias") == ':socer:'
-    assert emoji.emojize(':socer:', language="alias",
-                  variant="text_type") == ':socer:'
+    assert emoji.emojize(':tester:', language='alias') == ':tester:'
+    assert emoji.emojize(':footbal:', language='alias') == ':footbal:'
+    assert emoji.emojize(':socer:', language='alias') == ':socer:'
+    assert emoji.emojize(':socer:', language='alias', variant='text_type') == ':socer:'
 
 
-def test_demojize_name_only():
+def test_demojize_name_only(load_all_languages):  # type:ignore
     for emj, item in emoji.EMOJI_DATA.items():
         if item['status'] != emoji.STATUS['fully_qualified']:
             continue
         for lang_code in emoji.LANGUAGES:
-            if not lang_code in item:
+            if lang_code not in item:
                 continue
             name = item[lang_code]
             oneway = emoji.emojize(name, language=lang_code)
@@ -281,23 +385,41 @@ def test_demojize_complicated_string():
 
 
 def test_demojize_delimiters():
-    for e in ['\U000026BD', '\U0001f44d', '\U0001F3C8']:
-        for d in [(":", ":"), ("}", "}"), ("!$", "!!$"), ("[123", "456]"), ("😁", "👌"), ("[", "]")]:
+    for e in ['\U000026bd', '\U0001f44d', '\U0001f3c8']:
+        for d in [
+            (':', ':'),
+            ('}', '}'),
+            ('!$', '!!$'),
+            ('[123', '456]'),
+            ('😁', '👌'),
+            ('[', ']'),
+        ]:
             s = emoji.demojize(e, delimiters=d)
             assert s.startswith(d[0])
             assert s.endswith(d[1])
 
-    text = "Example with an emoji%sin a sentence and %s %s %s %s multiple emoji %s%s%s%s%s in a row"
-    for e in ['\U000026BD', '\U0001f44d', '\U0001F3C8']:
-        for d in [(":", ":"), ("{", "}"), ("!$", "$!"), (":", "::"), ("::", "::"), ("😁", "👌"), ("[", "]")]:
-            print("delimiter: %s" % (d, ))
-            text_with_unicode = text % ((e, ) * 10)
+    text = 'Example with an emoji%sin a sentence and %s %s %s %s multiple emoji %s%s%s%s%s in a row'
+    for e in ['\U000026bd', '\U0001f44d', '\U0001f3c8']:
+        for d in [
+            (':', ':'),
+            ('{', '}'),
+            ('!$', '$!'),
+            (':', '::'),
+            ('::', '::'),
+            ('😁', '👌'),
+            ('[', ']'),
+        ]:
+            print('delimiter: %s' % (d,))
+            text_with_unicode = text % ((e,) * 10)
             demojized_text = emoji.demojize(text_with_unicode, delimiters=d)
-            assert text_with_unicode != demojized_text, (text_with_unicode, demojized_text)
+            assert text_with_unicode != demojized_text, (
+                text_with_unicode,
+                demojized_text,
+            )
             assert e not in demojized_text
             assert emoji.emojize(demojized_text, delimiters=d) == text_with_unicode
             de = emoji.demojize(e, delimiters=d)
-            text_with_emoji = text % ((de, ) * 10)
+            text_with_emoji = text % ((de,) * 10)
             assert demojized_text == text_with_emoji
             assert emoji.emojize(text_with_emoji, delimiters=d) == text_with_unicode
 
@@ -305,11 +427,16 @@ def test_demojize_delimiters():
 def test_emoji_list():
     assert emoji.emoji_list('Hi, I am 👌 test')[0]['match_start'] == 9
     assert emoji.emoji_list('Hi') == []
-    if len('Hello 🇫🇷👌') < 10:  # skip these tests on python with UCS-2 as the string length/positions are different
+    if (
+        len('Hello 🇫🇷👌') < 10
+    ):  # skip these tests on python with UCS-2 as the string length/positions are different
         assert emoji.emoji_list('Hi, I am fine. 😁') == [
-            {'match_start': 15, 'match_end': 16, 'emoji': '😁'}]
+            {'match_start': 15, 'match_end': 16, 'emoji': '😁'}
+        ]
         assert emoji.emoji_list('Hello 🇫🇷👌') == [
-            {'emoji': '🇫🇷', 'match_start': 6, 'match_end': 8}, {'emoji': '👌', 'match_start': 8, 'match_end': 9}]
+            {'emoji': '🇫🇷', 'match_start': 6, 'match_end': 8},
+            {'emoji': '👌', 'match_start': 8, 'match_end': 9},
+        ]
 
 
 def test_distinct_emoji_list():
@@ -333,8 +460,9 @@ def test_replace_emoji():
     assert emoji.replace_emoji('Hello 🇫🇷👌', 'x') == 'Hello xx'
 
     def replace(emj: str, data: Dict[str, str]) -> str:
-        assert emj in ["🇫🇷", "👌"]
+        assert emj in ['🇫🇷', '👌']
         return 'x'
+
     assert emoji.replace_emoji('Hello 🇫🇷👌', replace) == 'Hello xx'
 
 
@@ -347,16 +475,39 @@ def test_is_emoji():
 
 
 def test_long_emoji():
-    assert emoji.demojize('This is \U0001F9D1\U0001F3FC\U0000200D\U0001F37C example text') == 'This is :person_feeding_baby_medium-light_skin_tone: example text'
-    assert emoji.demojize('This is \U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f468\U0001f3ff example text \U0001F469\U0001F3FB\U0000200D\U0001F91D\U0000200D\U0001F468\U0001F3FF') == 'This is :couple_with_heart_man_man_dark_skin_tone: example text :woman_and_man_holding_hands_light_skin_tone_dark_skin_tone:'
-    assert emoji.demojize('This is \U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f468\U0001f3ff\U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f48b\u200d\U0001f468\U0001f3ff example text \U0001F469\U0001F3FB\U0000200D\U0001F91D\U0000200D\U0001F468\U0001F3FF') == 'This is :couple_with_heart_man_man_dark_skin_tone::kiss_man_man_dark_skin_tone: example text :woman_and_man_holding_hands_light_skin_tone_dark_skin_tone:'
-    assert emoji.demojize('\U0001F46B\U0001F3FB This is \U0001f468\U0001f3ff\U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f468\U0001f3ff\U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f48b\u200d\U0001f468\U0001f3ff example text \U0001F469\U0001F3FB\U0000200D\U0001F91D\U0000200D\U0001F468\U0001F3FF') == ':woman_and_man_holding_hands_light_skin_tone: This is :man_dark_skin_tone::couple_with_heart_man_man_dark_skin_tone::kiss_man_man_dark_skin_tone: example text :woman_and_man_holding_hands_light_skin_tone_dark_skin_tone:'
-    assert emoji.demojize('\U0001F46B\U0001F3FB\U0001F46B\U0001F3FB\U0001F469\U0001F3FB\U0000200D\U0001F91D\U0000200D\U0001F468\U0001F3FF\U0001FAF1\U0001F3FD\U0001FAF1\U0001F3FD\U0000200D\U0001FAF2\U0001F3FF') == ':woman_and_man_holding_hands_light_skin_tone::woman_and_man_holding_hands_light_skin_tone::woman_and_man_holding_hands_light_skin_tone_dark_skin_tone::rightwards_hand_medium_skin_tone::handshake_medium_skin_tone_dark_skin_tone:'
-    s = ":crossed_fingers_medium-light_skin_tone::crossed_fingers::crossed_fingers_dark_skin_tone:"
+    assert (
+        emoji.demojize('This is \U0001f9d1\U0001f3fc\U0000200d\U0001f37c example text')
+        == 'This is :person_feeding_baby_medium-light_skin_tone: example text'
+    )
+    assert (
+        emoji.demojize(
+            'This is \U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f468\U0001f3ff example text \U0001f469\U0001f3fb\U0000200d\U0001f91d\U0000200d\U0001f468\U0001f3ff'
+        )
+        == 'This is :couple_with_heart_man_man_dark_skin_tone: example text :woman_and_man_holding_hands_light_skin_tone_dark_skin_tone:'
+    )
+    assert (
+        emoji.demojize(
+            'This is \U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f468\U0001f3ff\U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f48b\u200d\U0001f468\U0001f3ff example text \U0001f469\U0001f3fb\U0000200d\U0001f91d\U0000200d\U0001f468\U0001f3ff'
+        )
+        == 'This is :couple_with_heart_man_man_dark_skin_tone::kiss_man_man_dark_skin_tone: example text :woman_and_man_holding_hands_light_skin_tone_dark_skin_tone:'
+    )
+    assert (
+        emoji.demojize(
+            '\U0001f46b\U0001f3fb This is \U0001f468\U0001f3ff\U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f468\U0001f3ff\U0001f468\U0001f3ff\u200d\u2764\ufe0f\u200d\U0001f48b\u200d\U0001f468\U0001f3ff example text \U0001f469\U0001f3fb\U0000200d\U0001f91d\U0000200d\U0001f468\U0001f3ff'
+        )
+        == ':woman_and_man_holding_hands_light_skin_tone: This is :man_dark_skin_tone::couple_with_heart_man_man_dark_skin_tone::kiss_man_man_dark_skin_tone: example text :woman_and_man_holding_hands_light_skin_tone_dark_skin_tone:'
+    )
+    assert (
+        emoji.demojize(
+            '\U0001f46b\U0001f3fb\U0001f46b\U0001f3fb\U0001f469\U0001f3fb\U0000200d\U0001f91d\U0000200d\U0001f468\U0001f3ff\U0001faf1\U0001f3fd\U0001faf1\U0001f3fd\U0000200d\U0001faf2\U0001f3ff'
+        )
+        == ':woman_and_man_holding_hands_light_skin_tone::woman_and_man_holding_hands_light_skin_tone::woman_and_man_holding_hands_light_skin_tone_dark_skin_tone::rightwards_hand_medium_skin_tone::handshake_medium_skin_tone_dark_skin_tone:'
+    )
+    s = ':crossed_fingers_medium-light_skin_tone::crossed_fingers::crossed_fingers_dark_skin_tone:'
     assert emoji.demojize(emoji.demojize(s)) == s
 
 
-def test_untranslated():
+def test_untranslated(load_all_languages):  # type:ignore
     for item in emoji.EMOJI_DATA.values():
         if item['status'] != emoji.STATUS['fully_qualified']:
             continue
@@ -364,7 +515,11 @@ def test_untranslated():
             # untranslated
             value = emoji.emojize(item['en'], language='en')
             roundtrip = emoji.demojize(value, language='es')
-            assert roundtrip == value, '%s != %s (from %s)' % (ascii(roundtrip), ascii(value), item['en'])
+            assert roundtrip == value, '%s != %s (from %s)' % (
+                ascii(roundtrip),
+                ascii(value),
+                item['en'],
+            )
         else:
             # translated
             value = emoji.emojize(item['en'], language='en')
@@ -372,7 +527,7 @@ def test_untranslated():
             assert roundtrip == item['es'], '%s != %s' % (roundtrip, item['es'])
 
 
-def test_text():
+def test_text(load_all_languages):  # type:ignore
     emoji.config.demojize_keep_zwj = True  # Restore default config value
     emoji.config.replace_emoji_keep_zwj = False  # Restore default config value
 
@@ -401,12 +556,11 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
     def add_random_emoji(
         text: str,
         lst: List[Tuple[str, Dict[str, Any]]],
-        select: Callable[[Dict[str, Any]], Union[str, Literal[False]]] = default_select
+        select: Callable[[Dict[str, Any]], Union[str, Literal[False]]] = default_select,
     ) -> Tuple[str, str, List[str]]:
-
         emoji_list: List[str] = []
-        text_with_unicode = ""
-        text_with_placeholder = ""
+        text_with_unicode = ''
+        text_with_placeholder = ''
         for i in range(0, len(text), 10):
             while True:
                 emj, emj_data = random.choice(lst)
@@ -415,7 +569,7 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
                     break
 
             if UCS2:
-                j = text.find(" ", i, i + 10)
+                j = text.find(' ', i, i + 10)
                 if j == -1:
                     continue
             else:
@@ -423,11 +577,11 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
 
             text_with_unicode += text[i:j]
             text_with_unicode += emj
-            text_with_unicode += text[j:i + 10]
+            text_with_unicode += text[j : i + 10]
 
             text_with_placeholder += text[i:j]
             text_with_placeholder += placeholder
-            text_with_placeholder += text[j:i + 10]
+            text_with_placeholder += text[j : i + 10]
 
             emoji_list.append(emj)
 
@@ -437,10 +591,16 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
         return s.replace('\u200d', '').replace('\ufe0f', '')
 
     all_emoji_list = list(emoji.EMOJI_DATA.items())
-    qualified_emoji_list = [(emj, item) for emj, item in emoji.EMOJI_DATA.items() if item['status'] == emoji.STATUS['fully_qualified']]
+    qualified_emoji_list = [
+        (emj, item)
+        for emj, item in emoji.EMOJI_DATA.items()
+        if item['status'] == emoji.STATUS['fully_qualified']
+    ]
 
     # qualified emoji
-    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(text, qualified_emoji_list)
+    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(
+        text, qualified_emoji_list
+    )
     assert emoji.demojize(text_with_unicode) == text_with_placeholder
     assert emoji.emojize(text_with_placeholder) == text_with_unicode
     if not UCS2:
@@ -451,11 +611,13 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
 
     # qualified emoji from "es"
     def select_es(emj_data: Dict[str, Any]) -> Union[str, Literal[False]]:
-        return emj_data["es"] if "es" in emj_data else False
+        return emj_data['es'] if 'es' in emj_data else False
 
-    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(text, qualified_emoji_list, select=select_es)
-    assert emoji.demojize(text_with_unicode, language="es") == text_with_placeholder
-    assert emoji.emojize(text_with_placeholder, language="es") == text_with_unicode
+    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(
+        text, qualified_emoji_list, select=select_es
+    )
+    assert emoji.demojize(text_with_unicode, language='es') == text_with_placeholder
+    assert emoji.emojize(text_with_placeholder, language='es') == text_with_unicode
     if not UCS2:
         assert emoji.replace_emoji(text_with_unicode, '') == text
     assert set(emoji.distinct_emoji_list(text_with_unicode)) == set(emoji_list)
@@ -464,11 +626,13 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
 
     # qualified emoji from "alias"
     def select_alias(emj_data: Dict[str, Any]) -> Union[str, Literal[False]]:
-        return emj_data["alias"][0] if "alias" in emj_data else False
+        return emj_data['alias'][0] if 'alias' in emj_data else False
 
-    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(text, qualified_emoji_list, select=select_alias)
-    assert emoji.demojize(text_with_unicode, language="alias") == text_with_placeholder
-    assert emoji.emojize(text_with_placeholder, language="alias") == text_with_unicode
+    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(
+        text, qualified_emoji_list, select=select_alias
+    )
+    assert emoji.demojize(text_with_unicode, language='alias') == text_with_placeholder
+    assert emoji.emojize(text_with_placeholder, language='alias') == text_with_unicode
     if not UCS2:
         assert emoji.replace_emoji(text_with_unicode, '') == text
     assert set(emoji.distinct_emoji_list(text_with_unicode)) == set(emoji_list)
@@ -476,7 +640,9 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
         assert lis['emoji'] == emoji_list[i]
 
     # all emoji
-    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(text, all_emoji_list)
+    text_with_unicode, text_with_placeholder, emoji_list = add_random_emoji(
+        text, all_emoji_list
+    )
     assert emoji.demojize(text_with_unicode) == text_with_placeholder
     assert clean(emoji.emojize(text_with_placeholder)) == clean(text_with_unicode)
     if not UCS2:
@@ -486,32 +652,44 @@ Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deseru
         assert lis['emoji'] == emoji_list[i]
 
 
-def test_text_multiple_times():
+def test_text_multiple_times(load_all_languages: Any):
     # Run test_text() multiple times because it relies on a random text
     for _ in range(100):
-        test_text()
+        test_text(load_all_languages)
 
 
 def test_invalid_chars():
-    invalidchar = "\U0001F20F"
-    assert emoji.demojize(invalidchar) == invalidchar, "%r != %r" % (ascii(emoji.demojize(invalidchar)), ascii(invalidchar))
-    assert emoji.demojize(invalidchar) == invalidchar, "%r != %r" % (ascii(emoji.demojize(invalidchar)), ascii(invalidchar))
+    invalidchar = '\U0001f20f'
+    assert emoji.demojize(invalidchar) == invalidchar, '%r != %r' % (
+        ascii(emoji.demojize(invalidchar)),
+        ascii(invalidchar),
+    )
+    assert emoji.demojize(invalidchar) == invalidchar, '%r != %r' % (
+        ascii(emoji.demojize(invalidchar)),
+        ascii(invalidchar),
+    )
 
-    invalidchar = "u\2302 ⌂"
-    assert emoji.demojize(invalidchar) == invalidchar, "%r != %r" % (ascii(emoji.demojize(invalidchar)), ascii(invalidchar))
-    assert emoji.demojize(invalidchar) == invalidchar, "%r != %r" % (ascii(emoji.demojize(invalidchar)), ascii(invalidchar))
+    invalidchar = 'u\2302 ⌂'
+    assert emoji.demojize(invalidchar) == invalidchar, '%r != %r' % (
+        ascii(emoji.demojize(invalidchar)),
+        ascii(invalidchar),
+    )
+    assert emoji.demojize(invalidchar) == invalidchar, '%r != %r' % (
+        ascii(emoji.demojize(invalidchar)),
+        ascii(invalidchar),
+    )
 
 
 def test_combine_with_component():
-    text = "Example of a combined emoji%sin a sentence"
+    text = 'Example of a combined emoji%sin a sentence'
 
-    combined = emoji.emojize(text % ":woman_dark_skin_tone:")
-    separated = emoji.emojize(text % ":woman::dark_skin_tone:")
-    assert combined == separated,  "%r != %r" % (ascii(combined), ascii(separated))
+    combined = emoji.emojize(text % ':woman_dark_skin_tone:')
+    separated = emoji.emojize(text % ':woman::dark_skin_tone:')
+    assert combined == separated, '%r != %r' % (ascii(combined), ascii(separated))
 
-    combined = emoji.emojize(text % ":woman_dark_skin_tone_white_hair:")
-    separated = emoji.emojize(text % ":woman::dark_skin_tone:\u200d:white_hair:")
-    assert combined == separated,  "%r != %r" % (ascii(combined), ascii(separated))
+    combined = emoji.emojize(text % ':woman_dark_skin_tone_white_hair:')
+    separated = emoji.emojize(text % ':woman::dark_skin_tone:\u200d:white_hair:')
+    assert combined == separated, '%r != %r' % (ascii(combined), ascii(separated))
 
 
 purely_emoji_testdata = [
@@ -525,6 +703,6 @@ purely_emoji_testdata = [
 ]
 
 
-@pytest.mark.parametrize("string,expected", purely_emoji_testdata)
+@pytest.mark.parametrize('string,expected', purely_emoji_testdata)
 def test_purely_emoji(string: str, expected: bool):
     assert emoji.purely_emoji(string) == expected