+++ tests/test_dict.py
@@ -4,8 +4,10 @@ from typing import Set, Dict
 import re
 import emoji
 
+from testutils import load_all_languages as load_all_languages
 
-def test_all_languages_list():
+
+def test_all_languages_list(load_all_languages):  # type:ignore
     """Compare all language keys in EMOJI_DATA with the emoji.LANGUAGES list"""
 
     langs: Set[str] = set()
@@ -19,8 +21,8 @@ def test_all_languages_list():
 def test_emoji_versions():
     """Check that every emoji has a valid version"""
     for item in emoji.EMOJI_DATA.values():
-        assert "E" in item
-        v = item["E"]
+        assert 'E' in item
+        v = item['E']
         assert isinstance(v, (int, float))
         assert v >= 0.6
 
@@ -29,7 +31,7 @@ def check_duplicate_names(lang: str):
     """Check that there are no duplicate names in the fully_qualified except for different variants"""
     seen: Dict[str, int] = {}
     for item in emoji.EMOJI_DATA.values():
-        if item["status"] > emoji.STATUS["fully_qualified"]:
+        if item['status'] > emoji.STATUS['fully_qualified']:
             continue
 
         if lang not in item:
@@ -43,16 +45,16 @@ def check_duplicate_names(lang: str):
             seen[name] = 0
 
 
-def test_duplicate_names():
+def test_duplicate_names(load_all_languages):  # type:ignore
     """Check that there are no duplicate names in the fully_qualified except for different variants"""
     for lang in emoji.LANGUAGES:
         check_duplicate_names(lang)
 
 
-def test_name_valid():
+def test_name_valid(load_all_languages):  # type:ignore
     """Check that every name starts with colons and does not contain other colons or whitespace"""
 
-    pattern = re.compile(r":[^:\s]+:")
+    pattern = re.compile(r':[^:\s]+:')
     for item in emoji.EMOJI_DATA.values():
         for lang in emoji.LANGUAGES:
             if lang in item: